+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.1.1] - 2017-01-25
+- Added filter for payment source URL and description.
+- Added process_donation() method to make sure Pronamic gateway works correctly.
+
 ## [1.1.0] - 2016-11-08
 - Updated gateway system to Charitable version 1.3+.
 
@@ -36,7 +40,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-11-05
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/charitable/compare/1.1.0...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/charitable/compare/1.1.1...HEAD
+[1.1.1]: https://github.com/wp-pay-extensions/charitable/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/wp-pay-extensions/charitable/compare/1.0.5...1.1.0
 [1.0.5]: https://github.com/wp-pay-extensions/charitable/compare/1.0.4...1.0.5
 [1.0.4]: https://github.com/wp-pay-extensions/charitable/compare/1.0.3...1.0.4